+++ packages/Input/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- webkit-search-cancel-button style, removed Input NUMBER type [@tristanjasper](https://github.com/tristanjasper).
+
 ## [0.2.26] - 2020-04-29
 
 ### Added